Added updated station list; Added JetSetRadioFuture.Live
M3U and other files are now up-to-date with the site. Not a whole lot of songs have been added/removed from existing stations since this has been lasted updated. JetSetRadioFuture.Live also contains a large majority of the same songs as the original site, but has older stations that are no longer hosted